tools/libxl & tools/libxc: Add a new "-l" lazy switch.
Add a new "-l" option to xl restore which implies use of lazy restore.
Plumb the lazy argument from libxl to libxc.

Please note that the lazy restore is to be enabled only for HVM guests.
If "-l" is passed as an argument to non HVM guest then the process will
exit with error.

Signed-off-by: Dushyant Behl <myselfdushyantbehl@gmail.com>