tools/libxc: refactored mempaging code from xenpaging to libxc.
This patch moves the code to initialize and teardown mempaging from xenpaging
to libxc. The code refactored from xenpaging is the code which sets up paging,
initializes a shared ring and event channel to communicate with xen. This
communication is done between the hypervisor and the dom0 tool which performs
the activity of pager. Another routine is also added by this patch which will
teardown the ring and mem paging setup. The xenpaging code is changed to use
the newly created routines and is tested to properly build and work with this code.
The refactored code is to be placed in libxc with LGPL license while the old
xenpaging code was under GPL2+ license, hence proper permission has been taken
from previous copyright holders to relicense the code.

The refactoring is done so that the tools lazy restore pager and xenpaging which
rely on memory paging can use the same routines to initialize mempaging. This refactoring
will also allow any future (in-tree) tools to use mempaging. The refactored code in
xc_mem_paging_ring_setup is to be compiled into libxenguest.

Note:- The name of a macro in the file tools/ocaml/libs/xc/xenctrl_stubs.c is changed
from RING_SIZE to BUF_RING_SIZE because the earlier name collides with xen shared ring
definitions in io/ring.h

Signed-off-by: Dushyant Behl <myselfdushyantbehl@gmail.com>
Reviewed-by: Andres Lagar-Cavilla <andres@lagarcavilla.org>
Reviewed-by: Andrew Cooper <andrew.cooper3@citrix.com>
Acked-by: David Scott <dave.scott@citrix.com>